+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/utils/extensions/MessageListItem.kt
@@ -20,7 +20,7 @@ import com.getstream.sdk.chat.adapter.MessageListItem
 import io.getstream.chat.android.common.model.messsagelist.DateSeparatorState
 import io.getstream.chat.android.common.model.messsagelist.MessageItemState
 import io.getstream.chat.android.common.model.messsagelist.SystemMessageState
-import io.getstream.chat.android.common.model.messsagelist.ThreadSeparatorState
+import io.getstream.chat.android.common.model.messsagelist.ThreadDateSeparatorState
 import io.getstream.chat.android.common.model.messsagelist.TypingItemState
 import io.getstream.chat.android.common.model.messsagelist.MessageListItemState as MessageListItemCommon
 
@@ -34,7 +34,7 @@ public fun MessageListItemCommon.toUiMessageListItem(): MessageListItem {
     return when (this) {
         is DateSeparatorState -> MessageListItem.DateSeparatorItem(date = date)
         is SystemMessageState -> MessageListItem.MessageItem(message = message)
-        is ThreadSeparatorState -> MessageListItem.ThreadSeparatorItem(date = date, messageCount = replyCount)
+        is ThreadDateSeparatorState -> MessageListItem.ThreadSeparatorItem(date = date, messageCount = replyCount)
         is TypingItemState -> MessageListItem.TypingItem(users = typingUsers)
         is MessageItemState -> MessageListItem.MessageItem(
             message = message,